Fix #32 github action to build with meson
See also https://docs.github.com/en/actions/quickstart
This is just a start, more stuff should be added

+++ src/lib/ankerl/svector.h
@@ -164,7 +164,7 @@ struct storage : public header {
             throw std::bad_alloc();
         }
         mem += offset_to_data;
-        if (mem > std::numeric_limits<std::ptrdiff_t>::max()) {
+        if (static_cast<uint64_t>(mem) > static_cast<uint64_t>(std::numeric_limits<std::ptrdiff_t>::max())) {
             throw std::bad_alloc();
         }
